[AdminListBundle]: Add Form & Request property and Response function in AdminListEvent
fix annotations adminlistevent

Add change to UPGRADE

Fix null annotation

Add Response function AdminListEvent

+++ src/Kunstmaan/AdminListBundle/Controller/AdminListController.php
@@ -141,7 +141,16 @@ abstract class AdminListController extends Controller
 
             // Don't redirect to listing when coming from ajax request, needed for url chooser.
             if ($form->isValid() && !$request->isXmlHttpRequest()) {
-                $this->container->get('event_dispatcher')->dispatch(AdminListEvents::PRE_ADD, new AdminListEvent($helper));
+                $adminListEvent = new AdminListEvent($helper, $request, $form);
+                $this->container->get('event_dispatcher')->dispatch(
+                    AdminListEvents::PRE_ADD,
+                    $adminListEvent
+                );
+
+                // Check if Response is given
+                if ($adminListEvent->getResponse() instanceof Response) {
+                    return $adminListEvent->getResponse();
+                }
 
                 // Check if Sortable interface is implemented
                 if ($configurator instanceof SortableInterface) {
@@ -153,7 +162,16 @@ abstract class AdminListController extends Controller
 
                 $em->persist($helper);
                 $em->flush();
-                $this->container->get('event_dispatcher')->dispatch(AdminListEvents::POST_ADD, new AdminListEvent($helper));
+                $this->container->get('event_dispatcher')->dispatch(
+                    AdminListEvents::POST_ADD,
+                    $adminListEvent
+                );
+
+                // Check if Response is given
+                if ($adminListEvent->getResponse() instanceof Response) {
+                    return $adminListEvent->getResponse();
+                }
+
                 $indexUrl = $configurator->getIndexUrl();
 
                 return new RedirectResponse(
@@ -220,15 +238,40 @@ abstract class AdminListController extends Controller
 
             // Don't redirect to listing when coming from ajax request, needed for url chooser.
             if ($form->isValid() && !$request->isXmlHttpRequest()) {
-                $this->container->get('event_dispatcher')->dispatch(AdminListEvents::PRE_EDIT, new AdminListEvent($helper));
+                $adminListEvent = new AdminListEvent($helper, $request, $form);
+                $this->container->get('event_dispatcher')->dispatch(
+                    AdminListEvents::PRE_EDIT,
+                    $adminListEvent
+                );
+
+                // Check if Response is given
+                if ($adminListEvent->getResponse() instanceof Response) {
+                    return $adminListEvent->getResponse();
+                }
+
                 $em->persist($helper);
                 $em->flush();
-                $this->container->get('event_dispatcher')->dispatch(AdminListEvents::POST_EDIT, new AdminListEvent($helper));
+                $this->container->get('event_dispatcher')->dispatch(
+                    AdminListEvents::POST_EDIT,
+                    $adminListEvent
+                );
+
+                // Check if Response is given
+                if ($adminListEvent->getResponse() instanceof Response) {
+                    return $adminListEvent->getResponse();
+                }
+
                 $indexUrl = $configurator->getIndexUrl();
 
-                return new RedirectResponse(
-                    $this->generateUrl($indexUrl['path'], isset($indexUrl['params']) ? $indexUrl['params'] : array())
-                );
+                // Don't redirect to listing when coming from ajax request, needed for url chooser.
+                if (!$request->isXmlHttpRequest()) {
+                    return new RedirectResponse(
+                        $this->generateUrl(
+                            $indexUrl['path'],
+                            isset($indexUrl['params']) ? $indexUrl['params'] : array()
+                        )
+                    );
+                }
             }
         }
 
@@ -247,7 +290,6 @@ abstract class AdminListController extends Controller
             )
         );
     }
-
 
     protected function doViewAction(AbstractAdminListConfigurator $configurator, $entityId, Request $request)
     {
@@ -304,10 +346,28 @@ abstract class AdminListController extends Controller
 
         $indexUrl = $configurator->getIndexUrl();
         if ($request->isMethod('POST')) {
-            $this->container->get('event_dispatcher')->dispatch(AdminListEvents::PRE_DELETE, new AdminListEvent($helper));
+            $adminListEvent = new AdminListEvent($helper, $request);
+            $this->container->get('event_dispatcher')->dispatch(
+                AdminListEvents::PRE_DELETE,
+                $adminListEvent
+            );
+
+            // Check if Response is given
+            if ($adminListEvent->getResponse() instanceof Response) {
+                return $adminListEvent->getResponse();
+            }
+
             $em->remove($helper);
             $em->flush();
-            $this->container->get('event_dispatcher')->dispatch(AdminListEvents::POST_DELETE, new AdminListEvent($helper));
+            $this->container->get('event_dispatcher')->dispatch(
+                AdminListEvents::POST_DELETE,
+                $adminListEvent
+            );
+
+            // Check if Response is given
+            if ($adminListEvent->getResponse() instanceof Response) {
+                return $adminListEvent->getResponse();
+            }
         }
 
         return new RedirectResponse(
@@ -401,9 +461,9 @@ abstract class AdminListController extends Controller
     {
         // Check if Sortable interface is implemented
         if ($configurator instanceof SortableInterface) {
-            $route = function (EntityInterface $item) use ($configurator){
+            $route = function (EntityInterface $item) use ($configurator) {
                 return array(
-                    'path' => $configurator->getPathByConvention() . '_move_up',
+                    'path' => $configurator->getPathByConvention().'_move_up',
                     'params' => array('id' => $item->getId()),
                 );
             };
@@ -411,9 +471,9 @@ abstract class AdminListController extends Controller
             $action = new SimpleItemAction($route, 'arrow-up', 'Move up');
             $configurator->addItemAction($action);
 
-            $route = function (EntityInterface $item) use ($configurator){
+            $route = function (EntityInterface $item) use ($configurator) {
                 return array(
-                    'path' => $configurator->getPathByConvention() . '_move_down',
+                    'path' => $configurator->getPathByConvention().'_move_down',
                     'params' => array('id' => $item->getId()),
                 );
             };
